event-consumers/age-verification/requirements.txt
python-dotenv==1.0.0
kafka-python==2.0.2
redis==5.0.1
asyncio==3.4.3 